split to component
component로 정리
